pyproject.toml
[tool.poetry]
name = "KerasRec"
version = "0.1.0"
description = "A modular framework for news recommendation systems research using Keras and Hydra"
authors = ["Igor L.R. Azevedo <igorlima1740@gmail.com>", "David Phol <davidphol@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
tensorflow = "2.17.0"
hydra-core = "^1.3.2"
wandb = "^0.15.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
tqdm = "^4.65.0"
PyYAML = "^6.0"
plotly = "^5.13.0"
seaborn = "^0.12.0"
wordcloud = "^1.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.4.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
disallow_untyped_defs = true
